pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pinn4csi"
version = "0.1.0"
description = "Physics-Informed Neural Networks for WiFi CSI Analysis"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "PINN4CSI Research Team"},
]
license = {text = "MIT"}

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
]
notebook = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[tool.setuptools]
packages = ["pinn4csi", "pinn4csi.models", "pinn4csi.physics", "pinn4csi.data", "pinn4csi.training", "pinn4csi.utils", "pinn4csi.configs"]

[tool.setuptools.package-data]
pinn4csi = ["configs/**/*.yaml"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N", "B", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["pinn4csi"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "pinn4csi.*"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests as requiring GPU (deselect with '-m \"not gpu\"')",
]
addopts = "-m \"not slow and not gpu\""
